create instructions component and test
