+++ src/js/features/StartPage/StartPage.scss
@@ -17,27 +17,3 @@
     border-bottom-right-radius: 0;
   }
 }
-
-.instructions-list {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  font-size: 24px;
-  background-color: $white;
-  padding: 20px;
-  flex-grow: 1;
-
-  li {
-    display: flex;
-    align-items: center;
-    margin-bottom: 20px;
-
-    i {
-      margin-right: 10px;
-    }
-
-    &:last-child {
-      margin-bottom: 0;
-    }
-  }
-}
